Design: 산책중 ui 구현 완료(#2)
